pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "slide-builder"
version = "0.2.0"
description = "Cursor-native PowerPoint presentation generator"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "python-pptx>=0.6.23",
    "pydantic>=2.0",
    "pymupdf>=1.24",
    "python-docx>=1.1",
    "pyyaml>=6.0",
    "pdfplumber>=0.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
